Reorganize syntax tests
