+++ src/syntax/formats/charge/uc-charge.lexer.spec.ts
@@ -1,10 +1,4 @@
-import { beforeAll, beforeEach, describe, expect, it } from '@jest/globals';
-import { esline } from 'esgen';
-import { UcdCompiler } from '../../../compiler/deserialization/ucd-compiler.js';
-import { UC_MODULE_CHURI } from '../../../compiler/impl/uc-modules.js';
-import { ucMap } from '../../../schema/map/uc-map.js';
-import { UcDeserializer } from '../../../schema/uc-deserializer.js';
-import { ucUnknown } from '../../../schema/unknown/uc-unknown.js';
+import { beforeEach, describe, expect, it } from '@jest/globals';
 import {
   UC_TOKEN_AMPERSAND,
   UC_TOKEN_APOSTROPHE,
@@ -31,9 +25,7 @@ import {
   UC_TOKEN_SLASH,
   UcToken,
 } from '../../uc-token.js';
-import { UcURIParamsLexer } from '../uri-params/uc-uri-params.lexer.js';
 import { UcChargeLexer } from './uc-charge.lexer.js';
-import { ucFormatCharge } from './uc-inset-charge.js';
 
 describe('UcChargeLexer', () => {
   let lexer: UcChargeLexer;
@@ -228,70 +220,6 @@ describe('UcChargeLexer', () => {
         UC_TOKEN_PREFIX_SPACE,
         'abcdef',
       ]);
-    });
-  });
-});
-
-describe('ucInsetCharge', () => {
-  describe('in default mode', () => {
-    let readValue: UcDeserializer<unknown>;
-
-    beforeAll(async () => {
-      const compiler = new UcdCompiler({
-        models: {
-          readValue: {
-            model: ucMap({
-              a: ucUnknown({ within: { uriParam: ucFormatCharge() } }),
-            }),
-            lexer: ({ emit }) => {
-              const Lexer = UC_MODULE_CHURI.import(UcURIParamsLexer.name);
-
-              return esline`return new ${Lexer}(${emit})`;
-            },
-          },
-        },
-      });
-
-      ({ readValue } = await compiler.evaluate());
-    });
-
-    it('URI-decodes values', () => {
-      expect(readValue(`a='te%20st'`)).toEqual({ a: `te st'` });
-      expect(readValue(`a=te+st`)).toEqual({ a: 'te+st' });
-      expect(readValue(`a=%33`)).toEqual({ a: 3 });
-    });
-  });
-
-  describe('in plus-as-space mode', () => {
-    let readValue: UcDeserializer<unknown>;
-
-    beforeAll(async () => {
-      const compiler = new UcdCompiler({
-        models: {
-          readValue: {
-            model: ucMap({
-              a: ucUnknown({
-                within: {
-                  uriParam: ucFormatCharge({ plusAsSpace: true }),
-                },
-              }),
-            }),
-            lexer: ({ emit }) => {
-              const Lexer = UC_MODULE_CHURI.import(UcURIParamsLexer.name);
-
-              return esline`return new ${Lexer}(${emit})`;
-            },
-          },
-        },
-      });
-
-      ({ readValue } = await compiler.evaluate());
-    });
-
-    it('decodes URI charge values', () => {
-      expect(readValue(`a='te%20st'`)).toEqual({ a: `te st'` });
-      expect(readValue(`a='te+st'`)).toEqual({ a: `te st'` });
-      expect(readValue(`a=%33`)).toEqual({ a: 3 });
     });
   });
 });
